requirements.txt
autocommand==2.2.2
backports.tarfile==1.2.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
decorator==4.4.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
moviepy==1.0.3
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pip==23.2.1
platformdirs==4.2.2
proglog==0.1.10
python-dotenv==1.0.1
reportlab==4.2.5
requests==2.32.3
setuptools==75.6.0
tomli==2.0.1
tqdm==4.67.1
typeguard==4.3.0
typing-extensions==4.12.2
urllib3==2.2.3
wheel==0.43.0
zipp==3.19.2